Rename actual_authentication_backend to just authentication_backend.
