+++ frontend/src/components/contexts/contexts.ts
@@ -1,6 +1,5 @@
 import * as React from "react";
 import { AuthenticationHandlerService } from "../../services/authentication_handler_service/authentication_handler_service";
-import { AuthenticationBackendService } from "../../services/actual_backend/actual_authentication_backend";
 import { ProfileHandlerService } from "../../services/profile_handler_service/profile_handler_service";
 import { YourClubsHandlerService } from "../../services/your_clubs_handler_service/your_clubs_handler_service";
